Enum telamon_cuda::PerfCounter

pub enum PerfCounter {
    InstExecuted,
    ElapsedCyclesSM,
    LocalLoad,
    LocalStore,
    SharedLoad,
    SharedStore,
    GlobalLoadReplay,
    GlobalStoreReplay,
    SharedLoadReplay,
    SharedStoreReplay,
    UncachedGlobalLoadTransaction,
    GlobalStoreTransaction,
    L1LocalLoadHit,
    L1LocalLoadMiss,
    L1LocalStoreHit,
    L1LocalStoreMiss,
    L1GlobalLoadHit,
    L1GlobalLoadMiss,
    L1LocalSharedBankConflict,
    L2Subp0WriteSectorMisses,
    L2Subp1WriteSectorMisses,
    L2Subp0ReadSectorMisses,
    L2Subp1ReadSectorMisses,
    GldInst8Bit,
    GldInst16Bit,
    GldInst32Bit,
    GldInst64Bit,
    GldInst128Bit,
    GstInst8Bit,
    GstInst16Bit,
    GstInst32Bit,
    GstInst64Bit,
    GstInst128Bit,
    Fb0Subp0ReadSectors,
    Fb0Subp0WriteSectors,
    Fb1Subp0ReadSectors,
    Fb1Subp0WriteSectors,
    Fb0Subp1ReadSectors,
    Fb0Subp1WriteSectors,
    Fb1Subp1ReadSectors,
    Fb1Subp1WriteSectors,
    L2Subp0WriteSectorQueries,
    L2Subp1WriteSectorQueries,
    L2Subp0ReadSectorQueries,
    L2Subp1ReadSectorQueries,
    L2Subp0ReadTexSectorQueries,
    L2Subp1ReadTexSectorQueries,
    L2Subp0ReadHitSectors,
    L2Subp1ReadHitSectors,
    L2Subp0ReadTexHitSectors,
    L2Subp1ReadTexHitSectors,
    L2Subp0ReadSysmemSectorQueries,
    L2Subp1ReadSysmemSectorQueries,
    L2Subp0WriteSysmemSectorQueries,
    L2Subp1WriteSysmemSectorQueries,
    L2Subp0TotalReadSectorQueries,
    L2Subp1TotalReadSectorQueries,
    L2Subp0TotalWriteSectorQueries,
    L2Subp1TotalWriteSectorQueries,
    L2Subp2TotalReadSectorQueries,
    L2Subp2TotalWriteSectorQueries,
    L2Subp3TotalReadSectorQueries,
    L2Subp3TotalWriteSectorQueries,
    FbSubp0ReadSectors,
    FbSubp0WriteSectors,
    FbSubp1ReadSectors,
    FbSubp1WriteSectors,
    L2Subp2ReadSectorMisses,
    L2Subp3ReadSectorMisses,
    L2Subp2WriteSectorMisses,
    L2Subp3WriteSectorMisses,
    L2Subp0WriteL1SectorQueries,
    L2Subp1WriteL1SectorQueries,
    L2Subp2WriteL1SectorQueries,
    L2Subp3WriteL1SectorQueries,
    L2Subp0ReadL1SectorQueries,
    L2Subp1ReadL1SectorQueries,
    L2Subp2ReadL1SectorQueries,
    L2Subp3ReadL1SectorQueries,
    L2Subp0ReadL1HitSectors,
    L2Subp1ReadL1HitSectors,
    L2Subp2ReadL1HitSectors,
    L2Subp3ReadL1HitSectors,
}

Name a performance counter.

Variants
InstExecuted

The number of wrap of instruction executed, does not include replays.

ElapsedCyclesSM

The number of cycles used for the execution on each SMX.

LocalLoad

Loads and Stores

LocalStoreSharedLoadSharedStoreGlobalLoadReplay

Replays

GlobalStoreReplaySharedLoadReplaySharedStoreReplayUncachedGlobalLoadTransaction

Number of uncached global loads and global stores

GlobalStoreTransactionL1LocalLoadHit

The numbers for l1 cache accesses: hits and misses

L1LocalLoadMissL1LocalStoreHitL1LocalStoreMissL1GlobalLoadHitL1GlobalLoadMissL1LocalSharedBankConflictL2Subp0WriteSectorMissesL2Subp1WriteSectorMissesL2Subp0ReadSectorMissesL2Subp1ReadSectorMissesGldInst8BitGldInst16BitGldInst32BitGldInst64BitGldInst128BitGstInst8BitGstInst16BitGstInst32BitGstInst64BitGstInst128BitFb0Subp0ReadSectorsFb0Subp0WriteSectorsFb1Subp0ReadSectorsFb1Subp0WriteSectorsFb0Subp1ReadSectorsFb0Subp1WriteSectorsFb1Subp1ReadSectorsFb1Subp1WriteSectorsL2Subp0WriteSectorQueriesL2Subp1WriteSectorQueriesL2Subp0ReadSectorQueriesL2Subp1ReadSectorQueriesL2Subp0ReadTexSectorQueriesL2Subp1ReadTexSectorQueriesL2Subp0ReadHitSectorsL2Subp1ReadHitSectorsL2Subp0ReadTexHitSectorsL2Subp1ReadTexHitSectorsL2Subp0ReadSysmemSectorQueriesL2Subp1ReadSysmemSectorQueriesL2Subp0WriteSysmemSectorQueriesL2Subp1WriteSysmemSectorQueriesL2Subp0TotalReadSectorQueriesL2Subp1TotalReadSectorQueriesL2Subp0TotalWriteSectorQueriesL2Subp1TotalWriteSectorQueriesL2Subp2TotalReadSectorQueriesL2Subp2TotalWriteSectorQueriesL2Subp3TotalReadSectorQueriesL2Subp3TotalWriteSectorQueriesFbSubp0ReadSectorsFbSubp0WriteSectorsFbSubp1ReadSectorsFbSubp1WriteSectorsL2Subp2ReadSectorMissesL2Subp3ReadSectorMissesL2Subp2WriteSectorMissesL2Subp3WriteSectorMissesL2Subp0WriteL1SectorQueriesL2Subp1WriteL1SectorQueriesL2Subp2WriteL1SectorQueriesL2Subp3WriteL1SectorQueriesL2Subp0ReadL1SectorQueriesL2Subp1ReadL1SectorQueriesL2Subp2ReadL1SectorQueriesL2Subp3ReadL1SectorQueriesL2Subp0ReadL1HitSectorsL2Subp1ReadL1HitSectorsL2Subp2ReadL1HitSectorsL2Subp3ReadL1HitSectors